[#1554] Refactor Session model
